crawler/requirements.txt
httpx==0.26.0
beautifulsoup4==4.12.2
lxml==5.1.0
pymongo==4.6.1
python-dotenv==1.0.0
tenacity==8.2.3
python-docx==1.1.0
